+++ aldrin/src/low_level/call.rs
@@ -17,11 +17,11 @@ pub struct Call {
 }
 
 impl Call {
-    pub(crate) fn _new(client: Handle, serial: u32, function: u32, args: SerializedValue) -> Self {
+    pub(crate) fn new(client: Handle, serial: u32, function: u32, args: SerializedValue) -> Self {
         Self {
             function,
             args,
-            promise: Promise::_new(client, serial),
+            promise: Promise::new(client, serial),
         }
     }