+++ aldrin/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `PendingFunctionValue<T>`.
 - Add `Promise` and `low_level::Promise`.
 - Add `low_level::Call`.
+- Add `low_level::Service`.
 
 ### Changed